Refactor: Move cache initialization to cache module
This commit reorganizes the cache-related code by moving the initialization logic from `cache_config.py` to a new `cache.py` module. This improves code organization and makes the cache functionality more accessible.
