+++ public/static/assets/educational/characters/professor_pixel.png
