+++ public/static/assets/educational/environments/neotokyo_academy.png
