+++ src/ai_game_dev/cache.py
@@ -1,0 +1,5 @@
+"""Simple cache initialization for compatibility."""
+
+def initialize_sqlite_cache_and_memory():
+    """Placeholder for cache initialization - no longer using LangChain caching."""
+    pass
