--- src/ai_game_dev/examples/__init__.py
@@ -1,1 +1,0 @@
-"""Example game generation module integrated with Hatch build system."""
