--- src/ai_game_dev/main.py
@@ -1,37 +1,0 @@
-#!/usr/bin/env python3
-"""
-AI Game Development Package Demo
-Demonstrates the core functionality of the AI game development system
-"""
-
-import asyncio
-from pathlib import Path
-
-from ai_game_dev.library import AIGameDev
-from ai_game_dev.models import ComplexityLevel, GameSpec, GameType
-
-
-def main():
-    """Demo the AI game development system."""
-    print("🎮 AI Game Development System")
-    print("=" * 40)
-    
-    # Create a simple game specification
-    spec = GameSpec(
-        name="Demo Game",
-        description="A simple demo game generated by AI",
-        game_type=GameType.TWO_DIMENSIONAL,
-        features=["player movement", "basic gameplay"],
-        complexity=ComplexityLevel.BEGINNER
-    )
-    
-    print(f"✅ Game spec created: {spec.name}")
-    print(f"📋 Type: {spec.game_type}")
-    print(f"🎯 Complexity: {spec.complexity}")
-    print("🚀 AI Game Development system is ready!")
-    
-    print("=" * 40)
-
-
-if __name__ == "__main__":
-    main()
